Fix a bug related to cached loadPage and misused PromiseComponent:
The original way of using `PromiseComponent` is wrong in several aspects
and accidentally worked:

1. a) creating promise in `render()` resulted in redundant `fetch`
   events. b) However, it also forced `PromiseComponent` refresh
   the render result after every time it gets updated.

2. The `children` and `then` props of `PromiseComponent` should be
   mutually exclusive. If both presented, only `then` will be used.

However, the 0717f4d updates changed the behavior of `loadPage` and
furthermore cancelled the accidental correction effect by (1b).

This commit fixes the buggy usage of `PromiseComponent`. It combines
the two future events of `loadPage()` and status changing of `loading`
into one `Promise` object.